[VolvoOnCall] Backported fix for #8518 from 3.x
Changed trip ID data type to match API response.
Resolves #8518


[VolvoOnCall] Backported fix for #8554 from 3.x

User-Agent HTTP header is no longer sent in the requests
- as Jetty/9.4.20.v20190813 is blacklisted server-side.

The backported fix uses a shim wrapper for HttpUtil, allowing to inject
a customized Jetty's HttpClient, while keeping the rest of the interface
intact (goal was to minimize changes: 3.x branch has a more elegant fix
for the same, but coupled with other changes and refactorings that are
not part of this backport).
Resolves #8554


[VolvoOnCall] Backported HTTP req. cache from 3.x

This change is essential for not flooding VOC servers with requests on
each channel update. Cache idea and timeout configuration copied from 3.x
branch and adapted to 2.5.x codebase.
Related to #8554

Also-by: clinique <gael@lhopital.org>

+++ bundles/org.openhab.binding.volvooncall/src/main/java/org/openhab/binding/volvooncall/internal/handler/VehicleHandler.java
@@ -85,7 +85,7 @@ public class VehicleHandler extends BaseThingHandler {
     private VehiclePositionWrapper vehiclePosition = new VehiclePositionWrapper(new Position());
     private Status vehicleStatus = new Status();
     private @NonNullByDefault({}) VehicleConfiguration configuration;
-    private Integer lastTripId = 0;
+    private Long lastTripId = 0L; // Fix for #8518 (backported to 2.5.x)
 
     public VehicleHandler(Thing thing, VehicleStateDescriptionProvider stateDescriptionProvider) {
         super(thing);
@@ -131,7 +131,7 @@ public class VehicleHandler extends BaseThingHandler {
                         .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
 
                 if (thing.getProperties().containsKey(LAST_TRIP_ID)) {
-                    lastTripId = Integer.parseInt(thing.getProperties().get(LAST_TRIP_ID));
+                    lastTripId = Long.parseLong(thing.getProperties().get(LAST_TRIP_ID));
                 }
 
                 updateStatus(ThingStatus.ONLINE);
@@ -206,7 +206,7 @@ public class VehicleHandler extends BaseThingHandler {
             logger.debug("Trips discovered : {}", newTrips.size());
 
             if (!newTrips.isEmpty()) {
-                Integer newTripId = newTrips.get(newTrips.size() - 1).id;
+                Long newTripId = newTrips.get(newTrips.size() - 1).id;
                 if (newTripId > lastTripId) {
                     updateProperty(LAST_TRIP_ID, newTripId.toString());
                     lastTripId = newTripId;
